initial commit before testing

+++ contracts/Migrations.sol
@@ -11,7 +11,7 @@ pragma solidity ^0.8.6;
 */
 contract Migrations {
     address public owner;
-    uint public lastCompletedMigration;
+    uint256 public lastCompletedMigration;
 
     modifier restricted() {
         if (msg.sender == owner) _;
@@ -21,7 +21,7 @@ contract Migrations {
         owner = msg.sender;
     }
 
-    function setCompleted(uint completed) public restricted {
+    function setCompleted(uint256 completed) public restricted {
         lastCompletedMigration = completed;
     }
 }
